ticket_sniper/requirements.txt
uiautomator2>=3.0.0
pyyaml>=6.0
